pyproject.toml
[tool.poetry]
name = "matrix-madlib-bot"
version = "0.1.0"
description = "A bot for the Matrix protocol that generates madlibs."
authors = ["krazykirby99999"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
simplematrixbotlib = "^2.6.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
